requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudinary==1.28.0
crispy-bootstrap5==0.6
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.2
Django==3.2.9
django-allauth==0.51.0
django-cache-url==3.2.3
django-ckeditor==6.2.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-crispy-forms==1.14.0
django-ipware==4.0.2
django-js-asset==2.0.0
django-markdownx==3.0.1
django-mdeditor==0.1.20
django-pandas==0.6.6
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-recaptcha==2.0.7
environs==9.3.5
et-xmlfile==1.1.0
firebase-admin==5.2.0
frozenlist==1.2.0
future==0.18.2
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
line-bot-sdk==2.0.1
Markdown==3.3.6
marshmallow==3.14.1
metadata-parser==0.10.5
msgpack==1.0.4
multidict==5.2.0
numpy==1.21.4
oauthlib==3.2.0
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
proto-plus==1.20.5
protobuf==3.20.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==3.4.0
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
yarl==1.7.2
zipp==3.7.0
